some text animation with react spriing
