+++ package.json
@@ -23,6 +23,7 @@
     "next": "13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-waypoint": "^10.3.0",
     "splitting": "^1.0.6",
     "typescript": "5.2.2",
     "vanta": "^0.5.24"
