+++ pages/home/fSeries.tsx
