+++ pages/home/lineBanner.tsx
